requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
bcrypt==3.2.0
boto3==1.21.4
botocore==1.24.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.4
colorama==0.4.4
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==36.0.1
Django==4.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
Markdown==3.3.6
MarkupSafe==2.1.0
mysqlclient==2.1.0
openapi-codec==1.3.2
packaging==21.3
paramiko==2.9.2
Pillow==9.0.1
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
s3transfer==0.5.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0
